% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_dataset.R
\name{are.compatible}
\alias{are.compatible}
\title{Check two GROAN.NoisyDataSet for dimension compatibility}
\usage{
are.compatible(nds1, nds2, verbose = FALSE)
}
\arguments{
\item{nds1}{the first GROAN.NoisyDataSet to be tested}

\item{nds2}{the second GROAN.NoisyDataSet to be tested}

\item{verbose}{boolean, if TRUE the function prints messages detailing the comparison.}
}
\value{
TRUE if the passed GROAN.NoisyDataSet are dimensionally compatible, FALSE otherwise
}
\description{
This function verifies that the two passed GROAN.NoisyDataSet objects have
the same dimensions and can thus be used in the same experiment (typically training
models on one and testing on the other). The function returns a TRUE/FALSE. In \code{verbose}
mode the function also prints messages detailing the comparisons.
}
