% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificial_data.R
\name{artificialGenotypes}
\alias{artificialGenotypes}
\title{Create a set of artificial genotypes}
\usage{
artificialGenotypes(samples = 50, markers = 200, ploidy = 2)
}
\arguments{
\item{samples}{number of individuals (rows)}

\item{markers}{number of SNPs (columns)}

\item{ploidy}{number of alleles per SNP}
}
\value{
a samples x markers matrix
}
\description{
This function return a matrix of 0/1/2... up to ploidy level.
Nothing fancy, each data point is sampled from an equally
probable distribution.
}
\keyword{internal}
