% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{createRunId}
\alias{createRunId}
\title{Generate a random run id}
\usage{
createRunId()
}
\value{
a partially random alphanumeric string
}
\description{
This function returns a partially random alphanumeric string that can be
used to identify a single run.
}
