% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_dataset.R
\name{getExtraCovariatesNames}
\alias{getExtraCovariatesNames}
\title{String of extra covariates names}
\usage{
getExtraCovariatesNames(nds, separator = " ")
}
\arguments{
\item{nds}{a \code{Noisy Dataset} object.}

\item{separator}{used for string concatenation.}
}
\value{
a string representation of extra covariates names.
}
\description{
Given a \code{Noisy Dataset} object, this function
returns a representation of the extra covariates
present in the object. If no extra covariates are
present an empty string is returned.
}
\keyword{internal}
