% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_dataset.R
\name{getNoisyPhenotype}
\alias{getNoisyPhenotype}
\title{Generate an instance of noisy phenotypes}
\usage{
getNoisyPhenotype(nds)
}
\arguments{
\item{nds}{a \code{Noisy Dataset} object}
}
\value{
the phenotypes contained in \code{nds} with added noise.
}
\description{
Given a \code{Noisy Dataset} object, this function
applies the noise injector to the data and returns
a noisy version of it.
It is useful for inspecting the noisy injector effects.
}
