\name{groan.mu}
\alias{groan.mu}
\title{Specific growth rate (mu)}
\usage{
  groan.mu(x)
}
\arguments{
  \item{x}{A groan Curves object representing growth
  curves}
}
\value{
  A groan Curves object representing specific growth rate
  profiles
}
\description{
  Calculates the specific growth rate profile for a growth
  curve
}

