% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{GROAN.pea.SNPs}
\alias{GROAN.pea.SNPs}
\title{[DEPRECATED]}
\format{A data frame with 103 rows and 647 variables. Each
row represent a pea KI line, each column a SNP marker}
\source{
Annicchiarico et al., \emph{GBS-Based Genomic Selection
for Pea Grain Yield under Severe Terminal Drought}, The Plant Genome,
Volume 10. \doi{doi:10.3835/plantgenome2016.07.0072}
}
\usage{
GROAN.pea.SNPs
}
\description{
This piece of data is deprecated and will be dismissed in next
release. Please use \link{GROAN.KI} instead.
}
\keyword{datasets}
