% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{is.in.range}
\alias{is.in.range}
\title{Is in the passed numeric range}
\usage{
is.in.range(x, min = 0, max = 1)
}
\arguments{
\item{x}{value to be tested}

\item{min}{minimum admitted value}

\item{max}{maximum admitted value}
}
\value{
TRUE if x is a numeric completely in the specified range, FALSE otherwise
}
\description{
Function returns TRUE if the passed variable is numeric and all its content
is in the passed range (defined by \code{min} and \code{max}). Works also in
array (in that case all values must be in the range).
}
\keyword{internal}
