% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{is.naturalnumber}
\alias{is.naturalnumber}
\title{Is a positive integer?}
\usage{
is.naturalnumber(x, low.Limit = 0)
}
\arguments{
\item{x}{the value(s) to be tested}

\item{low.Limit}{the greatest value not accepted. Defaults to zero, meaning that
one is the smallest integer that returns true.}
}
\value{
TRUE if the value is a positive integer, FALSE otherwise (or NA)
}
\description{
\code{is.naturalnumber} returns true if the passed argument
is a positive integer, false otherwise.
Implementation taken from Marcog's answer
\href{http://stackoverflow.com/questions/4562257/what-is-the-fastest-way-to-check-if-a-number-is-a-positive-natural-number-in-r}{to this question}
}
\keyword{internal}
