% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{is.string}
\alias{is.string}
\title{Is a string?}
\usage{
is.string(x)
}
\arguments{
\item{x}{the variable to be checked}
}
\value{
a boolean
}
\description{
Returns TRUE if the passed x variable is a length one variable
containing characters values
}
\keyword{internal}
