\name{loess.xy.coords}
\alias{loess.xy.coords}
\title{Wrapper for \code{loess()} that uses \code{xy.coords}}
\usage{
  loess.xy.coords(x, y = NULL, ...)
}
\arguments{
  \item{x}{An \code{xy.coords} object, \code{list()} with
  named numeric elements \code{x} and \code{y}, or an
  atomic numeric vector of x-coordinates.  If an atomic
  vector the parameter \code{y} is required.}

  \item{y}{Required if the parameter \code{x} is specified
  as an atomic vector.  An atomic vector of y-coordinates.}

  \item{...}{Additional arguments to be passed to
  \code{loess()}}
}
\value{
  A list with components \describe{ \item{x}{x-coordinates
  of the smoothed curve} \item{y}{y-coordinates of the
  smoothed curve} \item{loess}{object returned by
  \code{loess()}} }
}
\description{
  Wrapper for \code{loess()} that uses \code{xy.coords}
}
\seealso{
  \link[stats]{loess}
}

