% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbench.R
\name{measurePredictionPerformance}
\alias{measurePredictionPerformance}
\title{Measure Performance of a Prediction}
\usage{
measurePredictionPerformance(truevals, predvals)
}
\arguments{
\item{truevals}{true values}

\item{predvals}{predicted values}
}
\value{
A named array with the following fields:
\describe{
 \item{pearson}{Pearson's correlation}
 \item{spearman}{Spearmans' correlation (order based)}
 \item{rmse}{Root Mean Square Error}
 \item{mae}{Mean Absolute Error}
 \item{coeff_det}{Coefficient of determination}
}
}
\description{
This method returns several performance metrics for the passed
predictions.
}
