\name{mu}
\alias{groan.mu}
\alias{mu}
\alias{specificRate}
\title{Specific growth rate (mu)}
\usage{
  mu(x)

  groan.mu(...)

  specificRate(...)
}
\arguments{
  \item{x}{A groan Curves object representing growth
  curves}
}
\value{
  A groan Curves object representing specific growth rate
  profiles
}
\description{
  Calculates the specific growth rate profile for a growth
  curve
}

