% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbench.R
\name{ndcg}
\alias{ndcg}
\title{Function to calculate mean Normalized Discounted Cumulative Gain (NDCG)}
\usage{
ndcg(y, y_hat, k = 0.2)
}
\arguments{
\item{y}{true values}

\item{y_hat}{predicted values}

\item{k}{relevant proportion of rank (top)}
}
\value{
a real value in [0,1]
}
\description{
This function calculates NDCG from the vectors of observed
and predicted values and the chosen proportion k of top
observations (rank).
}
