% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressors.R
\name{phenoRegressor.rrBLUP.G}
\alias{phenoRegressor.rrBLUP.G}
\title{NOT IMPLEMENTED - G BLUP using rrBLUP library}
\usage{
phenoRegressor.rrBLUP.G(phenotypes, covariances, extraCovariates = NULL, ...)
}
\arguments{
\item{phenotypes}{phenotypes, a numeric array (n x 1), missing values are predicted}

\item{covariances}{square matrix (n x n) of covariances.}

\item{extraCovariates}{optional extra covariates set, one row per phenotype (n), one column per covariate (w).
If NULL no extra covariates are considered.}

\item{...}{extra parameters are passed to rrBLUP::mixed.solve}
}
\value{
The function returns a list with the following fields:
\itemize{
  \item \code{predictions} : an array of (k) predicted phenotypes
  \item \code{hyperparams} : named vector with the following keys: Vu, Ve, beta, LL
  \item \code{extradata}   : list with information on trained model, coming from \code{\link[rrBLUP]{mixed.solve}}
}
}
\description{
This function SHOULD implement G blup using rrBLUP library. However I was
never able to make it work :( Not to be exported.
}
\keyword{internal}
