% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_dataset.R
\name{print.GROAN.NoisyDataset}
\alias{print.GROAN.NoisyDataset}
\title{Print a GROAN Noisy Dataset object}
\usage{
\method{print}{GROAN.NoisyDataset}(x, ...)
}
\arguments{
\item{x}{object of class GROAN.NoisyDataset.}

\item{...}{ignored, put here to match S3 function signature}
}
\value{
This function returns the original \code{GROAN.NoisyDataset} object invisibly (via \link[=invisible]{invisible(x)})
}
\description{
Short description for class GROAN.NoisyDataset, created with \link{createNoisyDataset}.
}
