% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbench.R
\name{print.GROAN.Workbench}
\alias{print.GROAN.Workbench}
\title{Print a GROAN Workbench object}
\usage{
\method{print}{GROAN.Workbench}(x, ...)
}
\arguments{
\item{x}{object of class GROAN.Workbench.}

\item{...}{ignored, put here to match S3 function signature}
}
\value{
This function returns the original \code{GROAN.Workbench} object invisibly (via \link[=invisible]{invisible(x)})
}
\description{
Short description for class GROAN.Workbench, created with \link{createWorkbench}.
}
