\name{smooth.adaptive.spline}
\alias{smooth.adaptive.spline}
\title{Adaptive smoothing using splines}
\usage{
  smooth.adaptive.spline(x, y = NULL, pct = 0.1,
    niter = NULL)
}
\arguments{
  \item{x}{An \code{xy.coords} object, \code{list()} with
  named numeric elements \code{x} and \code{y}, or an
  atomic numeric vector of x-coordinates.  If an atomic
  vector the parameter \code{y} is required.}

  \item{y}{Required if the parameter \code{x} is specified
  as an atomic vector.  An atomic vector of y-coordinates.}

  \item{pct}{Fraction of randomly selected data points
  excluded during each fitting iteration.  Specified on the
  interval [0,1].}

  \item{niter}{Number of fits to perform and aggregate into
  the final curve.  If not specified defaults to 10x more
  than the number of data points corresponding to
  \code{pct}.}
}
\value{
  A list with components \describe{ \item{x}{x-coordinates
  of the smoothed curve} \item{y}{y-coordinates of the
  smoothed curve} \item{pct}{fraction of data excluded}
  \item{niter}{number of fitting iterations performed} }
}
\description{
  Performs jack-knifed and bagged spline smoothing
}
\seealso{
  \link[stats]{smooth.spline} \link[stats]{spline}
}

