% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{stratified_crossvalidation_folds}
\alias{stratified_crossvalidation_folds}
\title{Assignments for stratified crossvalidation}
\usage{
stratified_crossvalidation_folds(strata, folds.num)
}
\arguments{
\item{strata}{an array of strata (will be treated as a factor)}

\item{folds.num}{the number of folds}
}
\value{
an array, same length as \code{strata}, of numbers in the 1:\code{folds.num} set
}
\description{
This function creates the assignments to perform a
stratified crossvalidation. It requires, for each
element of the set, a label describing the strata it
belongs to. It also requires the number of folds.
}
\details{
A warning is triggered if the number of folds is greater than the number of
elements of any stratum.
}
\keyword{internal}
