% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_dataset.R
\name{summary.GROAN.NoisyDataset}
\alias{summary.GROAN.NoisyDataset}
\title{Summary for GROAN Noisy Dataset object}
\usage{
\method{summary}{GROAN.NoisyDataset}(object, ...)
}
\arguments{
\item{object}{instance of class GROAN.NoisyDataset.}

\item{...}{additional arguments ignored, added for compatibility to generic \code{summary} function}
}
\value{
a data frame with GROAN.NoisyDataset stats.
}
\description{
Returns a dataframe with some description of an object created with \link{createNoisyDataset}.
}
