% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{summary.GROAN.Result}
\alias{summary.GROAN.Result}
\title{Summary of GROAN.Result}
\usage{
\method{summary}{GROAN.Result}(object, ...)
}
\arguments{
\item{object}{an object returned from \link{GROAN.run}}

\item{...}{additional arguments ignored, added for compatibility to generic \code{summary} function}
}
\value{
a data.frame with averaged statistics
}
\description{
Performance metrics are averaged over repetitions, so that a data.frame is produced
with one row per dataset/regressor/extra_covariates/strata/samples/markers/folds combination.
}
