% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{GROAN.pea.SNPs}
\alias{GROAN.pea.SNPs}
\title{Data - pea SNP genotypes}
\format{A data frame with 103 rows and 647 variables. Each
row represent a pea line, each column a SNP marker}
\source{
Annicchiarico et al., \emph{GBS-Based Genomic Selection
for Pea Grain Yield under Severe Terminal Drought}, The Plant Genome,
Volume 10. \doi{doi:10.3835/plantgenome2016.07.0072}
}
\usage{
GROAN.pea.SNPs
}
\description{
SNP genotypes for 103 pea lines. Each line of this data frame represent
a single pea line. Each column a SNP marker. Values can either be 0, 1, or 2,
representing the three possible genotypes (AA, Aa, and aa, respectively).
}
\keyword{datasets}
