% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{GROAN.pea.yield}
\alias{GROAN.pea.yield}
\title{Data - pea phenotypes (yield)}
\format{A named array with 103 slots.}
\source{
Annicchiarico et al., \emph{GBS-Based Genomic Selection
for Pea Grain Yield under Severe Terminal Drought}, The Plant Genome,
Volume 10. \doi{doi:10.3835/plantgenome2016.07.0072}
}
\usage{
GROAN.pea.yield
}
\description{
Phenotype dataset, containing data on pea grain yield [t/ha]
for each pea line.
}
\keyword{datasets}
