% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{is.boolean}
\alias{is.boolean}
\title{Is a boolean?}
\usage{
is.boolean(x)
}
\arguments{
\item{x}{the variable to be checked}
}
\value{
a boolean
}
\description{
Returns TRUE if the passed x variable is a length one variable
containing a valid TRUE/FALSE value. This test is stricter than
function \code{is.logical}, since NA, NULL and NaN all return FALSE.
Moreover, only single cell (lenght one) array admitted.
}
\keyword{internal}
